+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/elasticsearch_assets/ingest_processor_steps/index.ts
@@ -9,4 +9,3 @@ export { debugDeepCopyContextStep } from './debug_deep_copy_context_step';
 export { getDotExpanderSteps } from './get_dot_expander_steps';
 export { getRemoveEmptyFieldSteps } from './get_remove_empty_field_steps';
 export { removeEntityDefinitionFieldsStep } from './remove_entity_definition_fields_step';
-export { retentionDefinitionToIngestProcessorSteps } from './retention_definition_to_ingest_processor_steps';
