+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/entity_definitions/entity_descriptions/field_utils.ts
@@ -6,7 +6,7 @@
  */
 
 import type { MappingProperty } from '@elastic/elasticsearch/lib/api/types';
-import type { FieldDescription } from './types';
+import type { FieldDescription } from '../../installation/types';
 
 export const collectValues = ({
   destination,
