+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/entity_definitions/entity_descriptions/index.ts
@@ -5,6 +5,8 @@
  * 2.0.
  */
 
-export * from './get_united_definition';
-
-export { UnitedEntityDefinition } from './united_entity_definition';
+export * from './host';
+export * from './user';
+export * from './service';
+export * from './universal';
+export { getCommonFieldDescriptions } from './common';
