+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/entity_definitions/entity_descriptions/user.ts
@@ -5,8 +5,8 @@
  * 2.0.
  */
 import type { EntityDescription } from '../types';
-import { getCommonUnitedFieldDefinitions } from './common';
-import { collectValues as collect } from '../definition_utils';
+import { getCommonFieldDescriptions } from './common';
+import { collectValues as collect } from './field_utils';
 
 export const USER_DEFINITION_VERSION = '1.0.0';
 export const USER_IDENTITY_FIELD = 'user.name';
@@ -34,6 +34,6 @@ export const userEntityEngineDescription: EntityDescription = {
     collect({ source: 'user.hash' }),
     collect({ source: 'user.id' }),
     collect({ source: 'user.roles' }),
-    ...getCommonUnitedFieldDefinitions('user'),
+    ...getCommonFieldDescriptions('user'),
   ],
 };
