+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/entity_definitions/entity_manager_conversion.ts
@@ -1,0 +1,47 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { pick } from 'lodash/fp';
+import { entityDefinitionSchema, type EntityDefinition } from '@kbn/entities-schema';
+import { buildEntityDefinitionId } from '../utils';
+import type { EntityEngineInstallationDescriptor } from '../installation/types';
+
+export const convertToEntityManagerDefinition = (
+  description: EntityEngineInstallationDescriptor,
+  options: { fieldHistoryLength: number; indexPattern: string; namespace: string; filter: string }
+): EntityDefinition => {
+  const indexPatterns = options.indexPattern
+    ? description.indexPatterns.concat(options.indexPattern.split(','))
+    : description.indexPatterns;
+
+  const metadata = [
+    ...description.fields.map(pick(['source', 'destination', 'aggregation'])),
+    ...description.identityFields.map((source) => ({ source })),
+  ];
+
+  const definition = {
+    id: buildEntityDefinitionId(description.entityType, options.namespace),
+    name: `Security '${description.entityType}' Entity Store Definition`,
+    type: description.entityType,
+    indexPatterns,
+    identityFields: description.identityFields,
+    displayNameTemplate: `{{${description.identityFields.join('_')}}}`,
+    metadata,
+    latest: {
+      timestampField: description.settings.timestampField,
+      lookbackPeriod: description.settings.lookbackPeriod,
+      settings: {
+        syncDelay: description.settings.syncDelay,
+        frequency: description.settings.frequency,
+      },
+    },
+    version: description.version,
+    managed: true,
+  };
+
+  return entityDefinitionSchema.parse(definition);
+};
