+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/installation/engine_description.ts
@@ -4,27 +4,28 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-
 import { pipe } from 'fp-ts/lib/function';
 
 import { assign, concat, map, merge, update } from 'lodash/fp';
 import { set } from '@kbn/safer-lodash-set/fp';
-import type { EntityType } from '../../../../../common/api/entity_analytics';
-import type { EntityDescription, EntityEngineInstallationDescriptor } from './types';
-import { generateIndexMappings } from './united_entity_definition';
 
-import type { EntityStoreConfig } from '../types';
-import {
-  hostEntityEngineDescription,
-  userEntityEngineDescription,
-  universalEntityEngineDescription,
-} from './entity_types';
+import type { EntityType } from '../../../../../common/api/entity_analytics';
 import {
   DEFAULT_FIELD_HISTORY_LENGTH,
   DEFAULT_LOOKBACK_PERIOD,
   DEFAULT_TIMESTAMP_FIELD,
-} from './constants';
+} from '../entity_definitions/constants';
+import { generateIndexMappings } from '../elasticsearch_assets';
+import {
+  hostEntityEngineDescription,
+  userEntityEngineDescription,
+  universalEntityEngineDescription,
+} from '../entity_definitions/entity_descriptions';
+
+import type { EntityStoreConfig } from '../types';
 import { buildEntityDefinitionId } from '../utils';
+import type { EntityDescription } from '../entity_definitions/types';
+import type { EntityEngineInstallationDescriptor } from './types';
 
 const engineDescriptionRegistry: Record<EntityType, EntityDescription> = {
   host: hostEntityEngineDescription,
@@ -32,7 +33,6 @@ const engineDescriptionRegistry: Record<EntityType, EntityDescription> = {
   universal: universalEntityEngineDescription,
   service: getServiceUnitedDefinition,
 };
-
 export const getAvailableEntityTypes = () => Object.keys(engineDescriptionRegistry) as EntityType[];
 
 interface EngineDescriptionParams {
@@ -57,8 +57,8 @@ export const createEngineDescription = (options: EngineDescriptionParams) => {
   const description = engineDescriptionRegistry[entityType];
 
   const settings: EntityEngineInstallationDescriptor['settings'] = {
-    syncDelay: config.syncDelay,
-    frequency: config.frequency,
+    syncDelay: `${config.syncDelay.asSeconds()}s`,
+    frequency: `${config.frequency.asSeconds()}s`,
     lookbackPeriod: description.settings?.lookbackPeriod || DEFAULT_LOOKBACK_PERIOD,
     timestampField: description.settings?.timestampField || DEFAULT_TIMESTAMP_FIELD,
   };
@@ -66,8 +66,8 @@ export const createEngineDescription = (options: EngineDescriptionParams) => {
   const updatedDescription = pipe(
     description,
     set('id', buildEntityDefinitionId(entityType, namespace)),
-    update('indexPatterns', concat(indexPatterns)),
     update('settings', assign(settings)),
+    updateIndexPatterns(indexPatterns),
     updateRetentionFields(fieldHistoryLength),
     addIndexMappings
   ) as EntityEngineInstallationDescriptor;
@@ -75,16 +75,19 @@ export const createEngineDescription = (options: EngineDescriptionParams) => {
   return updatedDescription;
 };
 
+const updateIndexPatterns = (indexPatterns: string[]) =>
+  update('indexPatterns', (prev = []) => concat(indexPatterns, prev));
+
 const updateRetentionFields = (fieldHistoryLength: number) =>
   update(
     'fields',
     map(
       merge({
-        retention_operator: { maxLength: fieldHistoryLength },
+        retention: { maxLength: fieldHistoryLength },
         aggregation: { limit: fieldHistoryLength },
       })
     )
   );
 
 const addIndexMappings = (description: EntityEngineInstallationDescriptor) =>
-  set('indexMappings', description, generateIndexMappings(description));
+  set('indexMappings', generateIndexMappings(description), description);
