+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/installation/types.ts
@@ -6,35 +6,15 @@
  */
 
 import type {
+  MappingTypeMapping,
   IngestProcessorContainer,
   MappingProperty,
-  MappingTypeMapping,
 } from '@elastic/elasticsearch/lib/api/types';
+
 import type { EntityDefinition } from '@kbn/entities-schema';
 import type { EntityType } from '../../../../../common/api/entity_analytics';
-import type { EntityStoreConfig } from '../types';
 
-export type MappingProperties = NonNullable<MappingTypeMapping['properties']>;
 export type EntityDefinitionMetadataElement = NonNullable<EntityDefinition['metadata']>[number];
-
-type PickPartial<T, K extends keyof T, Optional extends K = never> = {
-  [P in K as P extends Optional ? never : P]: T[P];
-} & {
-  [P in K as P extends Optional ? P : never]?: Partial<T[P]>;
-};
-
-export type EntityDescription = PickPartial<
-  EntityEngineInstallationDescriptor,
-  | 'version'
-  | 'entityType'
-  | 'fields'
-  | 'identityFields'
-  | 'indexPatterns'
-  | 'indexMappings'
-  | 'settings'
-  | 'pipeline',
-  'indexPatterns' | 'indexMappings' | 'settings' | 'pipeline'
->;
 
 export interface EntityEngineInstallationDescriptor {
   id: string;
@@ -66,8 +46,8 @@ export interface EntityEngineInstallationDescriptor {
    * Any kibana.yml configuration will override these settings.
    */
   settings: {
-    syncDelay: EntityStoreConfig['syncDelay'];
-    frequency: EntityStoreConfig['frequency'];
+    syncDelay: string;
+    frequency: string;
     lookbackPeriod: string;
     timestampField: string;
   };
