+++ x-pack/plugins/security_solution/server/lib/entity_analytics/entity_store/task/field_retention_enrichment_task.ts
@@ -24,7 +24,7 @@ import {
 } from './state';
 import { INTERVAL, SCOPE, TIMEOUT, TYPE, VERSION } from './constants';
 import type { EntityAnalyticsRoutesDeps } from '../../types';
-import { getAvailableEntityTypes } from '../united_entity_definitions';
+
 import { executeFieldRetentionEnrichPolicy } from '../elasticsearch_assets';
 
 import { getEntitiesIndexName } from '../utils';
@@ -32,7 +32,8 @@ import {
   FIELD_RETENTION_ENRICH_POLICY_EXECUTION_EVENT,
   ENTITY_STORE_USAGE_EVENT,
 } from '../../../telemetry/event_based/events';
-import { VERSIONS_BY_ENTITY_TYPE } from '../united_entity_definitions/constants';
+import { VERSIONS_BY_ENTITY_TYPE } from '../entity_definitions/constants';
+import { getAvailableEntityTypes } from '../installation/engine_description';
 
 const logFactory =
   (logger: Logger, taskId: string) =>
